Move SearchMetadata to it's own file in utilities
